refactor: standardize codes (#22)
issues:
* refactor: standardize codes (#21)

commits:
* refactor: standardize codes (78d726c)
