+++ jest.config.ts
@@ -107,12 +107,12 @@ const config: Config = {
   // A map from regular expressions to module names or to arrays of module
   // names that allow to stub out resources with a single module
   moduleNameMapper: pathsToModuleNameMapper({
+    // @
+    '@': ['<rootDir>/src'],
+    '@/*': ['<rootDir>/src/*'],
     // @tests
     '@tests': ['<rootDir>/tests'],
-    '@tests/*': ['<rootDir>/tests/*'],
-    // @types
-    '@types': ['<rootDir>/src/types'],
-    '@types/*': ['<rootDir>/src/types/*']
+    '@tests/*': ['<rootDir>/tests/*']
   }),
 
   // An array of regexp pattern strings, matched against all module paths before
