+++ src/consts/index.ts
@@ -1,6 +1,20 @@
-export const PACKAGE_NAME = '@mnrendra/rollup-utils'
+import { normalize } from 'node:path'
 
-export const PLUGIN_PREFIX = '@mnrendra/rollup-plugin-'
+export const SCOPE = '@mnrendra'
+
+export const NAME = 'rollup-utils'
+
+export const PLUGIN_PREFIX = 'rollup-plugin-'
+
+export const DIST_NAME = `${SCOPE}/${NAME}`
+
+export const DIST_PLUGIN_PREFIX = `${SCOPE}/${PLUGIN_PREFIX}`
+
+export const DIST_PATH = `node_modules/${DIST_NAME}`
+
+export const DIST_PLUGIN_PATH_PREFIX = `node_modules/${DIST_PLUGIN_PREFIX}`
+
+export const SKIPPED_STACK = normalize(DIST_PATH)
 
 export const ANSI_COLORS = {
   RESET: '\x1b[0m',
