+++ tests/index.ts
@@ -1,9 +1,9 @@
+import * as dummies from './dummies'
 import * as mocks from './mocks'
 import * as unmocks from './unmocks'
-import * as stubs from './stubs'
 
 export {
+  dummies,
   mocks,
-  unmocks,
-  stubs
+  unmocks
 }
