+++ tsconfig.json
@@ -30,13 +30,12 @@
     "moduleResolution": "Bundler",                       /* Specify how TypeScript looks up a file from a given module specifier. */
     "baseUrl": "./src",                                  /* Specify the base directory to resolve non-relative module names. */
     "paths": {
+      // @
+      "@": ["./"],
+      "@/*": ["./*"],
       // @tests
       "@tests": ["../tests"],
-      "@tests/*": ["../tests/*"],
-      // @types
-      "@types": ["./types"],
-      "@types/*": ["./types/*"]
-
+      "@tests/*": ["../tests/*"]
     },                                                   /* Specify a set of entries that re-map imports to additional lookup locations. */
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
